rename use-tag-selector
